@drlaw1558 2025-02-04
Adding a quick note that I've done some initial testing on this too, with good results.  In the case of MIRI TSO observation jw03730013001_03101_00001-seg001_mirimage_uncal.fits the results are identical and jump step runtime goes down from 358 to 156 seconds.